A. 1.安裝npm; 2. 安裝 Laravel 的 ui 工具套件： composer require laravel/ui --dev 3. 建立含有使用者認證的bootstrap框架 php artisan ui bootstrap --auth 4. 安裝相依性套件並編譯 npm install && npm run dev npm run development B. 1. 編輯 app/User.php，加入三種帳號角色名稱的常數字串，並將 role 這個欄位加入 fillable 中。 2. 編輯 database/migrations/2014_10_12_000000_create_users_table.php 這一個建立使用者資料表的 migration 設定檔，在其中加入一欄用於儲存帳號角色的 role 欄位。 3. 編輯 app/Http/Controllers/Auth/RegisterController.php，設定使用者註冊時的角色都是一般使用者。 4. 建立 Gate 權限規則。 C. 自動生成三個 users: admin, manager, and user D. 更新兩個views： players.index 和teams.index，讓他們針對admin 與 manager 呈現受控管的超連結(新增、修改、刪除) E. 讓每一個路由都必須經過 auth 的 middleware 增加 can 的middleware，讓新增表單、儲存資料、刪除資料路由只能讓admin使用 不允許編輯表單給 user 權限使用，但允許 admin或manager使用

+++ routes/web.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\GamesController;
 use App\Http\Controllers\DevelopersController;
+use Illuminate\Support\Facades\Auth;
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -13,44 +14,53 @@ use App\Http\Controllers\DevelopersController;
 | contains the "web" middleware group. Now create something great!
 |
 */
+Route::middleware(['auth'])->group(function () {
+    
+    Route::get('/', function () {
+        return redirect('games');
+    });
+    
+    Route::get('games', [GamesController::class, 'index'])->name('games.index');
+    
+    Route::get('games/senior', [GamesController::class, 'senior'])->name('games.senior');
+    // 顯示單一球員資料
+    Route::get('games/{id}', [GamesController::class, 'show'])->where('id', '[0-9]+')->name('games.show');
+    // 修改單一球員表單
+    Route::get('games/{id}/edit', [GamesController::class, 'edit'])->where('id', '[0-9]+')->name('games.edit');
+    
+    Route::delete('games/delete/{id}', [GamesController::class, 'destroy'])->where('id', '[0-9]+')->name('games.destroy')->middleware('can:admin');
+    
+    Route::get('games/create', [GamesController::class, 'create'])->name('games.create')->middleware('can:admin')->middleware('can:admin');
+    // 修改球員表單
+    Route::get('games/{id}/edit', [GamesController::class, 'edit'])->where('id', '[0-9]+')->name('games.edit');
+    // 修改球員資料
+    Route::patch('games/update/{id}', [GamesController::class, 'update'])->where('id', '[0-9]+')->name('games.update');
+    // 儲存新球員資料
+    Route::post('games/store', [GamesController::class, 'store'])->where('id', '[0-9]+')->name('games.store')->middleware('can:admin');
+    
+    
+    
+    // 顯示顯示所有球隊資料
+    Route::get('developers', [DevelopersController::class, 'index'])->name('developers.index');
+    // 顯示單一球隊資料
+    Route::get('developers/{id}', [DevelopersController::class, 'show'])->where('id', '[0-9]+')->name('developers.show');
+    // 修改單一球隊表單
+    Route::get('developers/{id}/edit', [DevelopersController::class, 'edit'])->where('id', '[0-9]+')->name('developers.edit');
+    
+    Route::delete('developers/delete/{id}', [DevelopersController::class, 'destroy'])->where('id', '[0-9]+')->name('developers.destroy')->middleware('can:admin');
+    
+    Route::get('developers/create', [DevelopersController::class, 'create'])->name('developers.create')->middleware('can:admin');
+    // 修改球隊表單
+    Route::get('developers/{id}/edit', [DevelopersController::class, 'edit'])->where('id', '[0-9]+')->name('developers.edit');
+    // 修改球隊資料
+    Route::patch('developers/update/{id}', [DevelopersController::class, 'update'])->where('id', '[0-9]+')->name('developers.update');
+    // 儲存新球隊資料
+    Route::post('developers/store', [DevelopersController::class, 'store'])->name('developers.store')->middleware('can:admin');
 
-Route::get('/', function () {
-    return redirect('games');
 });
 
-Route::get('games', [GamesController::class, 'index'])->name('games.index');
 
-Route::get('games/senior', [GamesController::class, 'senior'])->name('games.senior');
-// 顯示單一球員資料
-Route::get('games/{id}', [GamesController::class, 'show'])->where('id', '[0-9]+')->name('games.show');
-// 修改單一球員表單
-Route::get('games/{id}/edit', [GamesController::class, 'edit'])->where('id', '[0-9]+')->name('games.edit');
+Auth::routes();
 
-Route::delete('games/delete/{id}', [GamesController::class, 'destroy'])->where('id', '[0-9]+')->name('games.destroy');
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
 
-Route::get('games/create', [GamesController::class, 'create'])->name('games.create');
-// 修改球員表單
-Route::get('games/{id}/edit', [GamesController::class, 'edit'])->where('id', '[0-9]+')->name('games.edit');
-// 修改球員資料
-Route::patch('games/update/{id}', [GamesController::class, 'update'])->where('id', '[0-9]+')->name('games.update');
-// 儲存新球員資料
-Route::post('games/store', [GamesController::class, 'store'])->where('id', '[0-9]+')->name('games.store');
-
-
-
-// 顯示顯示所有球隊資料
-Route::get('developers', [DevelopersController::class, 'index'])->name('developers.index');
-// 顯示單一球隊資料
-Route::get('developers/{id}', [DevelopersController::class, 'show'])->where('id', '[0-9]+')->name('developers.show');
-// 修改單一球隊表單
-Route::get('developers/{id}/edit', [DevelopersController::class, 'edit'])->where('id', '[0-9]+')->name('developers.edit');
-
-Route::delete('developers/delete/{id}', [DevelopersController::class, 'destroy'])->where('id', '[0-9]+')->name('developers.destroy');
-
-Route::get('developers/create', [DevelopersController::class, 'create'])->name('developers.create');
-// 修改球隊表單
-Route::get('developers/{id}/edit', [DevelopersController::class, 'edit'])->where('id', '[0-9]+')->name('developers.edit');
-// 修改球隊資料
-Route::patch('developers/update/{id}', [DevelopersController::class, 'update'])->where('id', '[0-9]+')->name('developers.update');
-// 儲存新球隊資料
-Route::post('developers/store', [DevelopersController::class, 'store'])->name('developers.store');
